Changes: 1. replace to collections GSC; 2. replace System log on cpw.mods.fml.common.FMLLog; 3. remove null check's, instanceof include null check; 4. random nextDouble() -> nextFloat(); 5. GL11.<...>d -> GL11.<...>f; 6. change BlockBase tickRate 0 -> 20 and other blocks tickRate to 20; 7. add AABB in TileEntityBase; 8. replace new Random() to XSTR; 9. association if's; 10. reduce getBlock() calls;

+++ src/main/java/lycanite/lycanitesmobs/api/entity/EntityCreatureAgeable.java
@@ -41,7 +41,7 @@ public abstract class EntityCreatureAgeable extends EntityCreatureBase {
     
     // ========== Setup ==========
     public void setupMob() {
-        if(this.babySpawnChance > 0D && this.rand.nextDouble() < this.babySpawnChance)
+        if(this.babySpawnChance > 0D && this.rand.nextFloat() < this.babySpawnChance)
         	this.setGrowingAge(growthTime);
         super.setupMob();
     }
