+++ src/main/java/lycanite/lycanitesmobs/api/spawning/SpawnTypeShadow.java
@@ -21,7 +21,7 @@ public class SpawnTypeShadow extends SpawnTypeBase {
     // ==================================================
     @Override
     public boolean canSpawn(long tick, World world, int x, int y, int z, boolean rare) {
-    	double roll = world.rand.nextDouble();
+    	double roll = world.rand.nextFloat();
     	ExtendedWorld worldExt = ExtendedWorld.getForWorld(world);
     	if(worldExt != null) {
     		if("shadowgames".equalsIgnoreCase(worldExt.getWorldEventType()))
