+++ src/main/java/lycanite/lycanitesmobs/api/spawning/SpawnTypeUndeath.java
@@ -34,7 +34,7 @@ public class SpawnTypeUndeath extends SpawnTypeDeath {
     // ==================================================
     @Override
     public boolean canSpawn(long tick, World world, int x, int y, int z, boolean rare) {
-    	double roll = world.rand.nextDouble();
+    	double roll = world.rand.nextFloat();
     	ExtendedWorld worldExt = ExtendedWorld.getForWorld(world);
     	if(worldExt != null) {
     		if("blackplague".equalsIgnoreCase(worldExt.getWorldEventType()))
