+++ src/main/java/lycanite/lycanitesmobs/arcticmobs/worldgen/WorldGenOozeLakes.java
@@ -67,7 +67,7 @@ public class WorldGenOozeLakes extends WorldGenLakes implements IWorldGenBase {
                 }
             }
 
-            if(typeValid && (this.generateSurfaceChance >= 1 || random.nextDouble() <= this.generateSurfaceChance)) {
+            if(typeValid && (this.generateSurfaceChance >= 1 || random.nextFloat() <= this.generateSurfaceChance)) {
                 int x = chunkX * 16 + random.nextInt(16);
                 int z = chunkZ * 16 + random.nextInt(16);
                 int y = random.nextInt(128);
@@ -75,7 +75,7 @@ public class WorldGenOozeLakes extends WorldGenLakes implements IWorldGenBase {
             }
         }
 
-        if(this.generateUndergroundChance > 0 && (this.generateUndergroundChance >= 1 || random.nextDouble() <= this.generateUndergroundChance)) {
+        if(this.generateUndergroundChance > 0 && (this.generateUndergroundChance >= 1 || random.nextFloat() <= this.generateUndergroundChance)) {
             int x = chunkX * 16 + random.nextInt(16);
             int z = chunkZ * 16 + random.nextInt(16);
             int top = Math.max(1, world.getTopSolidOrLiquidBlock(x, z) - 10);
